Move _extra_ modules to `extras/` and move `_testing.py` to this directory
