+++ .pre-commit-config.yaml
@@ -84,7 +84,7 @@ repos:
       - id: black-jupyter
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.1.9
+    rev: v0.1.11
     hooks:
       - id: ruff
         args: [ --fix, --exit-non-zero-on-fix ]
