+++ MANIFEST.in
@@ -31,9 +31,10 @@ prune docs/build
 # Misc
 include bin/tox
 recursive-include .github *.yaml
+recursive-include bin *.py
+recursive-include extras README.md *.py
 recursive-include misc *.py *.ipynb
 recursive-include requirements *.txt *.in *.md
-recursive-include bin *.py
 recursive-include tests *.py
 
 # Globally excluded patterns
