--- docs/api/internal/testing.rst
@@ -1,7 +1,0 @@
-Testing
-=======
-
-.. automodule:: ridgeplot._testing
-   :members:
-   :undoc-members:
-   :private-members:
