+++ extras/_ridgeplot_examples/_basic.py
