+++ extras/_ridgeplot_examples/_lincoln_weather.py
