+++ mypy.ini
@@ -1,7 +1,7 @@
 [mypy]
 
 # Import discovery  ---
-files = src/ridgeplot, tests, bin, docs
+files = src/ridgeplot, tests, bin, docs, extras
 namespace_packages = False
 
 # Disallow dynamic typing  ---
