+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 addopts =
     # Tests summary report
-    # ref: https://docs.pytest.org/en/7.2.x/how-to/output.html#producing-a-detailed-summary-report
+    # ref: https://docs.pytest.org/en/7.4.x/how-to/output.html#producing-a-detailed-summary-report
     -ra
     # (V)Verbose
     -vv
@@ -10,23 +10,27 @@ addopts =
     # Raise error on unknown markers
     --strict-markers
     # Import mode
-    # ref: https://docs.pytest.org/en/7.2.x/explanation/pythonpath.html#import-modes
+    # ref: https://docs.pytest.org/en/7.4.x/explanation/pythonpath.html#import-modes
     --import-mode=importlib
     # Fail if a test tries to open a network connection
     # ref: https://github.com/miketheman/pytest-socket
     --disable-socket
     # Coverage reports
     # ref: https://github.com/pytest-dev/pytest-cov
-    --cov
     --cov-report=term-missing
+    --cov-report=xml
+
+# List of directories that should be added to the python search path
+# https://docs.pytest.org/en/7.4.x/reference/reference.html#confval-pythonpath
+pythonpath = extras
 
 # Tests marked with @pytest.mark.xfail that actually
 # succeed will by default fail the test suite.
-# https://docs.pytest.org/en/7.2.x/reference/reference.html#confval-xfail_strict
+# https://docs.pytest.org/en/7.4.x/reference/reference.html#confval-xfail_strict
 xfail_strict = true
 
 # Warning filters pattern: action:message:category:module:line
-# ref: https://docs.pytest.org/en/7.2.x/how-to/capture-warnings.html
+# ref: https://docs.pytest.org/en/7.4.x/how-to/capture-warnings.html
 # ref: https://docs.python.org/3/library/warnings.html#describing-warning-filters
 # ref: https://docs.python.org/3/using/cmdline.html#cmdoption-W
 filterwarnings =
