+++ requirements/docs.in
@@ -1,10 +1,8 @@
 -e file:.
 
-beautifulsoup4
 furo
 jinja2
 kaleido
-minify-html
 myst-parser>=0.19.0
 sphinx
 sphinx-autobuild
@@ -20,3 +18,7 @@ sphinx-toolbox
 sphinx_design
 sphinx_thebe
 sphinxext-opengraph
+
+# extras/_compile_plotly_charts.py
+beautifulsoup4
+minify-html
