+++ requirements/locked/docs.38-darwin.txt
@@ -13,7 +13,7 @@ apeye==1.4.1
     # via sphinx-toolbox
 apeye-core==1.1.4
     # via apeye
-autodocsumm==0.2.11
+autodocsumm==0.2.12
     # via sphinx-toolbox
 babel==2.14.0
     # via sphinx
@@ -111,7 +111,7 @@ packaging==23.2
     #   statsmodels
 pandas==2.0.3
     # via statsmodels
-patsy==0.5.5
+patsy==0.5.6
     # via statsmodels
 platformdirs==4.1.0
     # via apeye
