+++ requirements/locked/local-dev.38-darwin.txt
@@ -57,7 +57,7 @@ pandas==2.0.3
     # via statsmodels
 parso==0.8.3
     # via jedi
-patsy==0.5.5
+patsy==0.5.6
     # via statsmodels
 pexpect==4.9.0
     # via ipython
@@ -112,7 +112,7 @@ tenacity==8.2.3
     # via plotly
 toposort==1.10
     # via pip-compile-multi
-traitlets==5.14.0
+traitlets==5.14.1
     # via
     #   ipython
     #   matplotlib-inline
@@ -122,7 +122,7 @@ typing-extensions==4.9.0
     #   ridgeplot
 tzdata==2023.4
     # via pandas
-wcwidth==0.2.12
+wcwidth==0.2.13
     # via prompt-toolkit
 wheel==0.42.0
     # via pip-tools
