+++ requirements/locked/mypy.38-darwin.txt
@@ -14,11 +14,11 @@ mypy-extensions==1.0.0
     # via mypy
 pandas-stubs==2.0.3.230814
     # via -r requirements/mypy.in
-types-beautifulsoup4==4.12.0.7
+types-beautifulsoup4==4.12.0.20240106
     # via -r requirements/mypy.in
-types-html5lib==1.1.11.15
+types-html5lib==1.1.11.20240106
     # via types-beautifulsoup4
-types-python-dateutil==2.8.19.14
+types-python-dateutil==2.8.19.20240106
     # via -r requirements/mypy.in
 types-pytz==2023.3.1.1
     # via
@@ -26,9 +26,9 @@ types-pytz==2023.3.1.1
     #   pandas-stubs
 types-pyyaml==6.0.12.12
     # via -r requirements/mypy.in
-types-requests==2.31.0.20231231
+types-requests==2.31.0.20240106
     # via -r requirements/mypy.in
-types-tqdm==4.66.0.5
+types-tqdm==4.66.0.20240106
     # via -r requirements/mypy.in
 urllib3==2.1.0
     # via types-requests
