+++ requirements/locked/release.38-darwin.txt
@@ -30,7 +30,7 @@ markdown-it-py==3.0.0
     # via rich
 mdurl==0.1.2
     # via markdown-it-py
-more-itertools==10.1.0
+more-itertools==10.2.0
     # via jaraco-classes
 nh3==0.2.15
     # via readme-renderer
