+++ requirements/locked/tests.38-darwin.txt
@@ -1,4 +1,4 @@
-# SHA1:2a51330c56a8d4fdae1a37be754fa92291959b4d
+# SHA1:206e21c36e23b580db9e92d08e812007dce48d21
 #
 # This file is autogenerated by pip-compile-multi
 # To update, run:
@@ -6,6 +6,8 @@
 #    pip-compile-multi
 #
 -e file:.
+    # via -r requirements/tests.in
+beautifulsoup4==4.12.2
     # via -r requirements/tests.in
 covdefaults==2.3.0
     # via -r requirements/tests.in
@@ -18,6 +20,8 @@ exceptiongroup==1.2.0
 iniconfig==2.0.0
     # via pytest
 kaleido==0.2.1
+    # via -r requirements/tests.in
+minify-html==0.15.0
     # via -r requirements/tests.in
 numpy==1.24.4
     # via
@@ -33,7 +37,7 @@ packaging==23.2
     #   statsmodels
 pandas==2.0.3
     # via statsmodels
-patsy==0.5.5
+patsy==0.5.6
     # via statsmodels
 plotly==5.18.0
     # via ridgeplot
@@ -58,6 +62,8 @@ six==1.16.0
     # via
     #   patsy
     #   python-dateutil
+soupsieve==2.5
+    # via beautifulsoup4
 statsmodels==0.14.1
     # via ridgeplot
 tenacity==8.2.3
