+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 labels =
     static = pre-commit, mypy
-    tests = unit-tests, e2e-tests, doctests
+    tests = tests-unit, tests-e2e, tests-extras, doctests
 isolated_build = true
 minversion = 3.20.0
 skipsdist = true
@@ -22,8 +22,9 @@ skip_install = true
 deps = -r requirements/locked/tests.{env:PY_PYTHON_VERSION}-{env:PY_SYS_PLATFORM}.txt
 commands =
     pytest: pytest {posargs:}
-    unit-tests: pytest tests/unit --cov-report=xml {posargs:}
-    e2e-tests: pytest tests/e2e {posargs:}
+    tests-unit: pytest tests/unit --cov {posargs:}
+    tests-e2e: pytest tests/e2e {posargs:}
+    tests-extras: pytest tests/extras {posargs:}
     doctests: pytest --doctest-modules src {posargs:}
 
 [testenv:pre-commit]
@@ -45,6 +46,10 @@ commands =
 
 [testenv:docs-{live,static}]
 description = generate Sphinx (live/static) HTML documentation
+setenv =
+    {[testenv]setenv}
+    # Append the extras/ directory to the PYTHONPATH
+    PYTHONPATH={env:PYTHONPATH}{:}{toxinidir}/extras
 deps = -r requirements/locked/docs.{env:PY_PYTHON_VERSION}-{env:PY_SYS_PLATFORM}.txt
 allowlist_externals = make
 commands =
